moved JSON writing to new Writer and JsonWriter, updated tests
