+++ simulariumio/constants.py
@@ -45,3 +45,7 @@ class DEFAULT_CAMERA_SETTINGS:
 
 
 DEFAULT_PLOT_MODE = "markers"
+
+BINARY_HEADER = "SIMULARIUMBIN"
+BINARY_VERSION = [1, 0, 0]
+BINARY_EOF = "\u005CEOFTHEFRAMEENDSHERE"
