+++ simulariumio/tests/converters/test_cytosim_converter.py
@@ -10,7 +10,7 @@ from simulariumio.cytosim import (
     CytosimObjectInfo,
     CytosimAgentInfo,
 )
-from simulariumio import MetaData
+from simulariumio import MetaData, JsonWriter
 from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 
 
@@ -1167,6 +1167,6 @@ from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 )
 def test_cytosim_converter(trajectory, expected_data):
     converter = CytosimConverter(trajectory)
-    buffer_data = converter._read_trajectory_data(converter._data)
+    buffer_data = JsonWriter.format_trajectory_data(converter._data)
     assert expected_data == buffer_data
     assert converter._check_agent_ids_are_unique_per_frame(buffer_data)
