+++ simulariumio/tests/converters/test_physicell_converter.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytest
 
 from simulariumio.physicell import PhysicellConverter, PhysicellData
-from simulariumio import MetaData
+from simulariumio import MetaData, JsonWriter
 from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 
 
@@ -204,6 +204,6 @@ from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 )
 def test_physicell_converter(trajectory, expected_data):
     converter = PhysicellConverter(trajectory)
-    buffer_data = converter._read_trajectory_data(converter._data)
+    buffer_data = JsonWriter.format_trajectory_data(converter._data)
     assert expected_data == buffer_data
     assert converter._check_agent_ids_are_unique_per_frame(buffer_data)
