+++ simulariumio/tests/converters/test_readdy_converter.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytest
 
 from simulariumio.readdy import ReaddyConverter, ReaddyData
-from simulariumio import UnitData, MetaData
+from simulariumio import UnitData, MetaData, JsonWriter
 from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 
 
@@ -225,6 +225,6 @@ from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 )
 def test_readdy_converter(trajectory, expected_data):
     converter = ReaddyConverter(trajectory)
-    buffer_data = converter._read_trajectory_data(converter._data)
+    buffer_data = JsonWriter.format_trajectory_data(converter._data)
     assert expected_data == buffer_data
     assert converter._check_agent_ids_are_unique_per_frame(buffer_data)
