+++ simulariumio/tests/converters/test_smoldyn_converter.py
@@ -8,7 +8,7 @@ from simulariumio.smoldyn import (
     SmoldynConverter,
     SmoldynData,
 )
-from simulariumio import MetaData, UnitData
+from simulariumio import MetaData, UnitData, JsonWriter
 from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 
 
@@ -391,6 +391,6 @@ from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 )
 def test_smoldyn_converter(trajectory, expected_data):
     converter = SmoldynConverter(trajectory)
-    buffer_data = converter._read_trajectory_data(converter._data)
+    buffer_data = JsonWriter.format_trajectory_data(converter._data)
     assert expected_data == buffer_data
     assert converter._check_agent_ids_are_unique_per_frame(buffer_data)
