+++ simulariumio/tests/converters/test_springsalad_converter.py
@@ -3,6 +3,7 @@
 
 import pytest
 
+from simulariumio import JsonWriter
 from simulariumio.springsalad import SpringsaladConverter, SpringsaladData
 from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 
@@ -208,6 +209,6 @@ from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 )
 def test_springsalad_converter(trajectory, expected_data):
     converter = SpringsaladConverter(trajectory)
-    buffer_data = converter._read_trajectory_data(converter._data)
+    buffer_data = JsonWriter.format_trajectory_data(converter._data)
     assert expected_data == buffer_data
     assert converter._check_agent_ids_are_unique_per_frame(buffer_data)
