+++ simulariumio/tests/filters/test_every_nth_subpoint_filter.py
@@ -3,7 +3,7 @@
 
 import pytest
 
-from simulariumio import FileConverter
+from simulariumio import FileConverter, JsonWriter
 from simulariumio.filters import EveryNthSubpointFilter
 from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 
@@ -741,6 +741,6 @@ from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
 def test_every_nth_subpoint_filter(input_path, _filter, expected_data):
     converter = FileConverter(input_path)
     filtered_data = converter.filter_data([_filter])
-    buffer_data = converter._read_trajectory_data(filtered_data)
+    buffer_data = JsonWriter.format_trajectory_data(filtered_data)
     assert expected_data == buffer_data
     assert converter._check_agent_ids_are_unique_per_frame(buffer_data)
