+++ simulariumio/tests/filters/test_multiply_time_filter.py
@@ -3,7 +3,7 @@
 
 import pytest
 
-from simulariumio import FileConverter
+from simulariumio import FileConverter, JsonWriter
 from simulariumio.filters import MultiplyTimeFilter
 from simulariumio.tests.conftest import test_scatter_plot
 from simulariumio.constants import DEFAULT_CAMERA_SETTINGS
@@ -303,6 +303,6 @@ def test_multiply_time_filter(input_path, plot_data, _filter, expected_data):
     converter = FileConverter(input_path)
     converter.add_plot(plot_data, "scatter")
     filtered_data = converter.filter_data([_filter])
-    buffer_data = converter._read_trajectory_data(filtered_data)
+    buffer_data = JsonWriter.format_trajectory_data(filtered_data)
     assert expected_data == buffer_data
     assert converter._check_agent_ids_are_unique_per_frame(buffer_data)
