+++ simulariumio/writers/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .json_writer import JsonWriter  # noqa: F401
